requirements.txt
tensorflow>=2.9.0
numpy
matplotlib
Pillow
transformers
torch